Add camera and fix steamid to ulong

+++ Mayday.Game/Networking/Packets/PlayerMovePacket.cs
@@ -5,7 +5,7 @@ namespace Mayday.Game.Networking.Packets
     public class PlayerMovePacket : INetworkPacket
     {
         
-        public uint SteamId { get; set; }
+        public ulong SteamId { get; set; }
         public int X { get; set; }
         public int Y { get; set; }
